Add 2 more drive motors

+++ include/motor.h
@@ -7,7 +7,7 @@ void smartMotorSlewEnabled(bool flag);
 
 short smartMotorGet(unsigned char channel);
 void smartMotorSet(unsigned char channel, short speed);
-void smartMotorSlew(unsigned char channel, short up, short down);
+void smartMotorSlew(unsigned char channel, float up, float down);
 void smartMotorReversed(unsigned char channel, bool reversed);
 
 #endif
